refactor(workflow): phase-1 review battery findings
Review battery (code-reviewer, silent-failure-hunter, thermo-nuclear audit)
findings applied:

Silent failures fixed:
- fission spec destroy now deletes Workflows (all three destroy paths);
  previously a spec-created Workflow survived destroy with exit 0.
- spec list / spec validate conflict-check / resource summary now include
  Workflows instead of silently omitting them.
- The spec parse path applies the same function-type defaulting as kubectl
  and workflow create, so spec validate/apply accepts the RFC's worked
  example verbatim.
- A Task function reference of type function-weights (an HTTPTrigger canary
  concern the engine cannot execute) is rejected at admission.
- Multi-document manifests are rejected explicitly: sigs.k8s.io/yaml decodes
  only the first document, which silently dropped the rest.
- workflow validate/graph guard the nil client on cluster-optional runs
  (previously a panic), validate checks functions in the manifest's
  namespace (not the flag fallback) and validates the name when present,
  and graph refuses empty specs and --file+--name together.
- The Workflow webhook warns on Fission.*-prefixed catch errorTypes outside
  the built-in set (a typo'd built-in is a silently dead route).

Structure:
- Workflow validation moved to its own workflow_validation.go (validation.go
  was already past the size bar).
- Shared RetryPolicy backoff-bounds check extracted; async keeps its
  MaxAsyncAttempts clamp, workflows their MaxWorkflowAttempts.
- State names pinned to ^[A-Za-z0-9_-]{1,64}$ before they become durable
  identifiers in run history.
- ValidateForAdmission identity wrappers dropped; webhooks call Validate.
